+++ R/RRRR_class.R
@@ -14,7 +14,7 @@ print.RRRR <- function(x,  digits = max(3L, getOption("digits") - 2L), ...){
   cat("Specifications:\n")
   print(do.call(base::c, x$spec))
   cat("\nCoefficients:\n")
-  print.default(coef(x), digits = digits)
+  print(coef(x), digits = digits)
 }
 
 #' Plot Objective value of a Robust Reduced-Rank Regression